Kokako -> SDDP
