+++ Project.toml
@@ -1,4 +1,4 @@
-name = "Kokako"
+name = "SDDP"
 uuid = "f4570300-c277-11e8-125c-4912f86ce65d"
 authors = ["Oscar Dowson <oscar.dowson@northwestern.edu"]
 version = "0.0.0"
