+++ docs/src/assets/spaghetti_plot.html
@@ -1,5 +1,5 @@
 <!--
-    Copyright 2018, Oscar Dowson.
+    Copyright 2017-19, Oscar Dowson.
     This Source Code Form is subject to the terms of the Mozilla Public License,
     v. 2.0. If a copy of the MPL was not distributed with this file, You can
     obtain one at http://mozilla.org/MPL/2.0/.
@@ -31,7 +31,7 @@ if(t==e.dx){for((r||f>e.dy)&&(f=e.dy);++u<o;)i=n[u],i.x=a,i.y=c,i.dy=f,a+=i.dx=M
 
     </script>
     <script>
-        //  Copyright 2018, Oscar Dowson.
+        //  Copyright 2017-19, Oscar Dowson.
 //  This Source Code Form is subject to the terms of the Mozilla Public License,
 //  v. 2.0. If a copy of the MPL was not distributed with this file, You can
 //  obtain one at http://mozilla.org/MPL/2.0/.
