+++ docs/src/tutorial/03_objective_uncertainty.md
@@ -1,5 +1,5 @@
 ```@meta
-CurrentModule = Kokako
+CurrentModule = SDDP
 ```
 
 # Basic III: objective uncertainty
@@ -29,19 +29,19 @@ In stage `t`, the objective is not to minimize
 ## Creating a model
 
 To add an uncertain objective, we can simply call [`@stageobjective`](@ref) from
-inside the [`Kokako.parameterize`](@ref) function.
+inside the [`SDDP.parameterize`](@ref) function.
 
 ```jldoctest tutorial_two
-using Kokako, GLPK
+using SDDP, GLPK
 
-model = Kokako.LinearPolicyGraph(
+model = SDDP.LinearPolicyGraph(
             stages = 3,
             sense = :Min,
             lower_bound = 0.0,
             optimizer = with_optimizer(GLPK.Optimizer)
         ) do subproblem, t
     # Define the state variable.
-    @variable(subproblem, 0 <= volume <= 200, Kokako.State, initial_value = 200)
+    @variable(subproblem, 0 <= volume <= 200, SDDP.State, initial_value = 200)
     # Define the control variables.
     @variables(subproblem, begin
         thermal_generation >= 0
@@ -61,7 +61,7 @@ model = Kokako.LinearPolicyGraph(
         (inflow = 50.0, fuel_multiplier = 1.0),
         (inflow = 100.0, fuel_multiplier = 0.75)
     ]
-    Kokako.parameterize(subproblem, Ω, [1/3, 1/3, 1/3]) do ω
+    SDDP.parameterize(subproblem, Ω, [1/3, 1/3, 1/3]) do ω
         JuMP.fix(inflow, ω.inflow)
         @stageobjective(subproblem,
             ω.fuel_multiplier * fuel_cost[t] * thermal_generation)
@@ -78,9 +78,9 @@ A policy graph with 3 nodes.
 
 As in the previous two tutorials, we train the policy:
 ```jldoctest tutorial_two; filter=r"Confidence interval.+?\n"
-Kokako.train(model; iteration_limit = 10)
+SDDP.train(model; iteration_limit = 10)
 
-simulations = Kokako.simulate(model, 500)
+simulations = SDDP.simulate(model, 500)
 
 objective_values = [
     sum(stage[:stage_objective] for stage in sim) for sim in simulations
@@ -92,7 +92,7 @@ using Statistics
 ci = round(1.96 * std(objective_values) / sqrt(500), digits = 2)
 
 println("Confidence interval: ", μ, " ± ", ci)
-println("Lower bound: ", round(Kokako.calculate_bound(model), digits = 2))
+println("Lower bound: ", round(SDDP.calculate_bound(model), digits = 2))
 
 # output
