+++ examples/StochDynamicProgramming.jl_stock.jl
@@ -1,4 +1,4 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -18,31 +18,31 @@
 
 =#
 
-using Kokako, GLPK, Test
+using SDDP, GLPK, Test
 
 function stock_example()
-    model = Kokako.PolicyGraph(Kokako.LinearGraph(5),
-                bellman_function = Kokako.AverageCut(lower_bound = -2),
+    model = SDDP.PolicyGraph(SDDP.LinearGraph(5),
+                bellman_function = SDDP.AverageCut(lower_bound = -2),
                 optimizer = with_optimizer(GLPK.Optimizer)
                                     ) do sp, stage
 
         # @state(sp, 0 <= state <= 1, state0 == 0.5)
-        @variable(sp, 0 <= state <= 1, Kokako.State, initial_value = 0.5)
+        @variable(sp, 0 <= state <= 1, SDDP.State, initial_value = 0.5)
         @variable(sp, 0 <= control <= 0.5)
         @variable(sp, ξ)
         @constraint(sp, state.out == state.in - control + ξ)
-        Kokako.parameterize(sp, 0.0:1/30:0.3) do ω
+        SDDP.parameterize(sp, 0.0:1/30:0.3) do ω
             JuMP.fix(ξ, ω)
         end
         @stageobjective(sp, (sin(3 * stage) - 1) * control)
     end
 
-    Kokako.train(model, iteration_limit = 50, print_level = 0)
-    @test Kokako.calculate_bound(model) ≈ -1.471 atol=0.001
+    SDDP.train(model, iteration_limit = 50, print_level = 0)
+    @test SDDP.calculate_bound(model) ≈ -1.471 atol=0.001
 
-    simulation_results = Kokako.simulate(model, 1_000)
+    simulation_results = SDDP.simulate(model, 1_000)
     @test length(simulation_results) == 1_000
-    @test Kokako.Statistics.mean(
+    @test SDDP.Statistics.mean(
         sum(data[:stage_objective] for data in simulation)
         for simulation in simulation_results
     ) ≈ -1.471 atol=0.05
