+++ examples/StructDualDynProg.jl_prob5.2_3stages.jl
@@ -1,4 +1,4 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -8,11 +8,11 @@
         https://github.com/blegat/StochasticDualDynamicProgramming.jl/blob/fe5ef82db6befd7c8f11c023a639098ecb85737d/test/prob5.2_3stages.jl
 ==#
 
-using Kokako, GLPK, Test
+using SDDP, GLPK, Test
 
 function test_prob52_3stages()
-    model = Kokako.PolicyGraph(Kokako.LinearGraph(3),
-                bellman_function = Kokako.AverageCut(lower_bound = 0.0),
+    model = SDDP.PolicyGraph(SDDP.LinearGraph(3),
+                bellman_function = SDDP.AverageCut(lower_bound = 0.0),
                 optimizer = with_optimizer(GLPK.Optimizer)
                 ) do sp, t
         n = 4
@@ -22,7 +22,7 @@ function test_prob52_3stages()
         T = [8760, 7000, 1500] / 8760
         D2 = [diff([0, 3919, 7329, 10315])  diff([0, 7086, 9004, 11169])]
         p2 = [0.9, 0.1]
-        @variable(sp, x[i=1:n] >= 0, Kokako.State, initial_value = 0.0)
+        @variable(sp, x[i=1:n] >= 0, SDDP.State, initial_value = 0.0)
         @variables(sp, begin
             y[1:n, 1:m] >= 0
             v[1:n]      >= 0
@@ -36,7 +36,7 @@ function test_prob52_3stages()
         end)
         @stageobjective(sp, ic'v + C' * y * T + 1e6 * penalty)
         if t != 1 # no uncertainty in first stage
-            Kokako.parameterize(sp, 1:size(D2, 2), p2) do ω
+            SDDP.parameterize(sp, 1:size(D2, 2), p2) do ω
                 for j in 1:m
                     JuMP.fix(ξ[j], D2[j, ω])
                 end
@@ -46,8 +46,8 @@ function test_prob52_3stages()
             @constraint(sp, sum(v) == 0)
         end
     end
-    Kokako.train(model, iteration_limit = 30, print_level = 0)
-    @test Kokako.calculate_bound(model) ≈ 406712.49 atol = 0.1
+    SDDP.train(model, iteration_limit = 30, print_level = 0)
+    @test SDDP.calculate_bound(model) ≈ 406712.49 atol = 0.1
     # sim = simulate(mod, 1, [:x, :penalty])
     # @test length(sim) == 1
     # @test isapprox(sim[1][:x][1], [2986,0,7329,854])
