+++ examples/infinite_horizon/ball_on_beam.jl
@@ -1,19 +1,19 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 # This should kind of work, but it doesn't.
-using Kokako, Ipopt, Test
+using SDDP, Ipopt, Test
 
 function infinite_ball_on_beam()
-    graph = Kokako.Graph(
+    graph = SDDP.Graph(
         :root_node,
         [:time_step],
         [(:root_node => :time_step, 1.0), (:time_step => :time_step, 0.999)]
     )
-    model = Kokako.PolicyGraph(graph,
-                bellman_function = Kokako.AverageCut(lower_bound = 0.0),
+    model = SDDP.PolicyGraph(graph,
+                bellman_function = SDDP.AverageCut(lower_bound = 0.0),
                 optimizer = with_optimizer(Ipopt.Optimizer, print_level = 0),
                 direct_mode = false,
                 sense = :Min
@@ -25,13 +25,13 @@ function infinite_ball_on_beam()
         τ = 1  #3  # Maximum torque.
         @variables(subproblem, begin
             # Beam position.
-            r, Kokako.State, (initial_value = 1)
+            r, SDDP.State, (initial_value = 1)
             # Rate of change of beam position.
-            r′, Kokako.State, (initial_value = 0)
+            r′, SDDP.State, (initial_value = 0)
             # Angle of beam.
-            θ, Kokako.State, (initial_value = -0.1745)
+            θ, SDDP.State, (initial_value = -0.1745)
             # Rate of change of angle of beam.
-            θ′, Kokako.State, (initial_value = 0)
+            θ′, SDDP.State, (initial_value = 0)
 
             # Control variable: torque to apply to beam.
             -τ <= u <= τ
@@ -57,13 +57,13 @@ using Random
 Random.seed!(1234)
 model = infinite_ball_on_beam()
 
-Kokako.train(model, iteration_limit = 2_000, print_level = 1,
+SDDP.train(model, iteration_limit = 2_000, print_level = 1,
     cycle_discretization_delta = 0.001)
 
 maximum_depth = 0
 for depth in 10:10:100
     try
-        simulation = Kokako.simulate(model;
+        simulation = SDDP.simulate(model;
             max_depth = depth, terminate_on_cycle = false)
         global maximum_depth = depth
     catch
@@ -72,7 +72,7 @@ for depth in 10:10:100
     end
 end
 
-simulation = Kokako.simulate(model, 1, [:r, :θ, :u],
+simulation = SDDP.simulate(model, 1, [:r, :θ, :u],
     max_depth = maximum_depth, terminate_on_cycle = false)
 using Plots
 plot(
