+++ src/visualization/publication_plot.jl
@@ -1,9 +1,9 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-# Internal function: convert dataset (from Kokako.simulate) into a matrix where
+# Internal function: convert dataset (from SDDP.simulate) into a matrix where
 # the rows are quantiles, and the columns are stages.
 function publication_data(dataset::Vector{Vector{Dict{Symbol, Any}}},
         quantiles::Vector{Float64}, stage_function::Function)
@@ -17,7 +17,7 @@ function publication_data(dataset::Vector{Vector{Dict{Symbol, Any}}},
 end
 
 """
-    Kokako.publication_plot(
+    SDDP.publication_plot(
         data_function, simulations;
         quantile = [0.0, 0.1, 0.25, 0.5, 0.75, 0.9, 1.0],
         kwargs...)
@@ -28,7 +28,7 @@ See `Plots.jl` for the list of keyword arguments.
 
 ### Example
 
-    Kokako.publication_plot(simulations; title = "My title") do data
+    SDDP.publication_plot(simulations; title = "My title") do data
         return data[:stage_objective]
     end
 """
@@ -37,7 +37,7 @@ function publication_plot(
         kwargs...)
     # An annoying over-load so that we can provide a consistent interface
     # instead of the Plots.jl generated `publicationplot`.
-    return Kokako.publicationplot(simulations, data_function, kwargs...)
+    return SDDP.publicationplot(simulations, data_function, kwargs...)
 end
 
 RecipesBase.@userplot PublicationPlot
