+++ src/visualization/spaghetti_plot.html
@@ -1,5 +1,5 @@
 <!--
-    Copyright 2018, Oscar Dowson.
+    Copyright 2017-19, Oscar Dowson.
     This Source Code Form is subject to the terms of the Mozilla Public License,
     v. 2.0. If a copy of the MPL was not distributed with this file, You can
     obtain one at http://mozilla.org/MPL/2.0/.
