+++ test/plugins/risk_measures.jl
@@ -1,16 +1,16 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-using Kokako, Test
+using SDDP, Test
 
 using GLPK  # Required for Wasserstein.
 
 @testset "Expectation" begin
     risk_adjusted_probability = Vector{Float64}(undef, 5)
-    Kokako.adjust_probability(
-        Kokako.Expectation(),
+    SDDP.adjust_probability(
+        SDDP.Expectation(),
         risk_adjusted_probability,
         [0.1, 0.2, 0.3, 0.4, 0.5],
         [:a, :b, :c, :d, :e],
@@ -19,8 +19,8 @@ using GLPK  # Required for Wasserstein.
     @test risk_adjusted_probability == [0.1, 0.2, 0.3, 0.4, 0.5]
 
     risk_adjusted_probability = Vector{Float64}(undef, 5)
-    Kokako.adjust_probability(
-        Kokako.Expectation(),
+    SDDP.adjust_probability(
+        SDDP.Expectation(),
         risk_adjusted_probability,
         [0.1, 0.2, 0.3, 0.3, 0.1],
         [:a, :b, :c, :d, :e],
@@ -31,8 +31,8 @@ end
 
 @testset "WorstCase" begin
     risk_adjusted_probability = Vector{Float64}(undef, 5)
-    Kokako.adjust_probability(
-        Kokako.WorstCase(),
+    SDDP.adjust_probability(
+        SDDP.WorstCase(),
         risk_adjusted_probability,
         [0.1, 0.2, 0.0, 0.4, 0.5],
         [:a, :b, :c, :d, :e],
@@ -41,8 +41,8 @@ end
     @test risk_adjusted_probability == [1.0, 0.0, 0.0, 0.0, 0.0]
 
     risk_adjusted_probability = Vector{Float64}(undef, 5)
-    Kokako.adjust_probability(
-        Kokako.WorstCase(),
+    SDDP.adjust_probability(
+        SDDP.WorstCase(),
         risk_adjusted_probability,
         [0.1, 0.2, 0.3, 0.4, 0.5],
         [:a, :b, :c, :d, :e],
@@ -52,26 +52,26 @@ end
 end
 
 @testset "Constructors" begin
-   a = Kokako.Expectation()
-   b = Kokako.AVaR(0.5)
-   c = Kokako.WorstCase()
+   a = SDDP.Expectation()
+   b = SDDP.AVaR(0.5)
+   c = SDDP.WorstCase()
    d = 0.5a + 0.3b + 0.2c
    @test d.measures[1] == (0.5, a)
    @test d.measures[2] == (0.3, b)
    @test d.measures[3] == (0.2, c)
 
-   aa = Kokako.EAVaR(lambda=0.5, beta=0.25)
-   @test aa.measures[1] == (0.5, Kokako.Expectation())
-   @test aa.measures[2] == (0.5, Kokako.AVaR(0.25))
+   aa = SDDP.EAVaR(lambda=0.5, beta=0.25)
+   @test aa.measures[1] == (0.5, SDDP.Expectation())
+   @test aa.measures[2] == (0.5, SDDP.AVaR(0.25))
 end
 
 @testset "AV@R" begin
-    @test_throws Exception Kokako.AVaR(-0.1)
-    @test_throws Exception Kokako.AVaR(1.1)
+    @test_throws Exception SDDP.AVaR(-0.1)
+    @test_throws Exception SDDP.AVaR(1.1)
     @testset "beta=0.2" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
-            Kokako.AVaR(0.2),
+        SDDP.adjust_probability(
+            SDDP.AVaR(0.2),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
             [:a, :b, :c, :d],
@@ -81,8 +81,8 @@ end
     end
     @testset "beta=0" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
-            Kokako.AVaR(0.0),
+        SDDP.adjust_probability(
+            SDDP.AVaR(0.0),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
             [:a, :b, :c, :d],
@@ -92,8 +92,8 @@ end
     end
     @testset "beta=1" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
-            Kokako.AVaR(1.0),
+        SDDP.adjust_probability(
+            SDDP.AVaR(1.0),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
             [:a, :b, :c, :d],
@@ -104,15 +104,15 @@ end
 end
 
 @testset "EAV@R" begin
-    @test_throws Exception Kokako.EAVaR(lambda=1.1)
-    @test_throws Exception Kokako.EAVaR(lambda=-0.1)
-    @test_throws Exception Kokako.EAVaR(beta=1.1)
-    @test_throws Exception Kokako.EAVaR(beta=-0.1)
+    @test_throws Exception SDDP.EAVaR(lambda=1.1)
+    @test_throws Exception SDDP.EAVaR(lambda=-0.1)
+    @test_throws Exception SDDP.EAVaR(beta=1.1)
+    @test_throws Exception SDDP.EAVaR(beta=-0.1)
     @testset "Max - (0.25, 0.2)" begin
         nominal_probability = [0.1, 0.2, 0.3, 0.4]
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
-            Kokako.EAVaR(lambda=0.25, beta=0.2),
+        SDDP.adjust_probability(
+            SDDP.EAVaR(lambda=0.25, beta=0.2),
             risk_adjusted_probability,
             nominal_probability,
             [:a, :b, :c, :d],
@@ -124,8 +124,8 @@ end
     @testset "Min - (0.25, 0.2)" begin
         nominal_probability = [0.1, 0.2, 0.3, 0.4]
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
-            Kokako.EAVaR(lambda=0.25, beta=0.2),
+        SDDP.adjust_probability(
+            SDDP.EAVaR(lambda=0.25, beta=0.2),
             risk_adjusted_probability,
             nominal_probability,
             [:a, :b, :c, :d],
@@ -137,8 +137,8 @@ end
     @testset "Max - (0.5, 0.0)" begin
         nominal_probability = [0.1, 0.2, 0.3, 0.4]
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
-            Kokako.EAVaR(lambda=0.5, beta=0.0),
+        SDDP.adjust_probability(
+            SDDP.EAVaR(lambda=0.5, beta=0.0),
             risk_adjusted_probability,
             nominal_probability,
             [:a, :b, :c, :d],
@@ -150,8 +150,8 @@ end
     @testset "Max - (0.5, 0.0) 2" begin
         nominal_probability = [0.0, 0.2, 0.4, 0.4]
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
-            Kokako.EAVaR(lambda=0.5, beta=0.0),
+        SDDP.adjust_probability(
+            SDDP.EAVaR(lambda=0.5, beta=0.0),
             risk_adjusted_probability,
             nominal_probability,
             [:a, :b, :c, :d],
@@ -165,8 +165,8 @@ end
 @testset "ModifiedChiSquared" begin
     @testset "Min - R=0.25" begin
         risk_adjusted_probability = Vector{Float64}(undef, 5)
-        Kokako.adjust_probability(
-            Kokako.ModifiedChiSquared(0.25),
+        SDDP.adjust_probability(
+            SDDP.ModifiedChiSquared(0.25),
             risk_adjusted_probability,
             fill(0.2, 5),
             [:a, :b, :c, :d, :e],
@@ -177,8 +177,8 @@ end
     end
     @testset "Max - R=0.25" begin
         risk_adjusted_probability = Vector{Float64}(undef, 5)
-        Kokako.adjust_probability(
-            Kokako.ModifiedChiSquared(0.25),
+        SDDP.adjust_probability(
+            SDDP.ModifiedChiSquared(0.25),
             risk_adjusted_probability,
             fill(0.2, 5),
             [:a, :b, :c, :d, :e],
@@ -189,8 +189,8 @@ end
     end
     @testset "Min - R=0.4" begin
         risk_adjusted_probability = Vector{Float64}(undef, 5)
-        Kokako.adjust_probability(
-            Kokako.ModifiedChiSquared(0.4),
+        SDDP.adjust_probability(
+            SDDP.ModifiedChiSquared(0.4),
             risk_adjusted_probability,
             fill(0.2, 5),
             [:a, :b, :c, :d, :e],
@@ -201,8 +201,8 @@ end
     end
     @testset "Max - R=0.4" begin
         risk_adjusted_probability = Vector{Float64}(undef, 5)
-        Kokako.adjust_probability(
-            Kokako.ModifiedChiSquared(0.4),
+        SDDP.adjust_probability(
+            SDDP.ModifiedChiSquared(0.4),
             risk_adjusted_probability,
             fill(0.2, 5),
             [:a, :b, :c, :d, :e],
@@ -213,8 +213,8 @@ end
     end
     @testset "Min - R=√0.8" begin
         risk_adjusted_probability = Vector{Float64}(undef, 5)
-        Kokako.adjust_probability(
-            Kokako.ModifiedChiSquared(sqrt(0.8)),
+        SDDP.adjust_probability(
+            SDDP.ModifiedChiSquared(sqrt(0.8)),
             risk_adjusted_probability,
             fill(0.2, 5),
             [:a, :b, :c, :d, :e],
@@ -226,14 +226,14 @@ end
 
 @testset "Wasserstein" begin
     function default_wasserstein(alpha)
-        return Kokako.Wasserstein(with_optimizer(GLPK.Optimizer); alpha = alpha) do x, y
+        return SDDP.Wasserstein(with_optimizer(GLPK.Optimizer); alpha = alpha) do x, y
             return abs(x - y)
         end
     end
     @test_throws Exception default_wasserstein(-1.0)
     @testset ":Max Worst case" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
+        SDDP.adjust_probability(
             default_wasserstein(10.0),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
@@ -244,7 +244,7 @@ end
     end
     @testset ":Min Worst case" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
+        SDDP.adjust_probability(
             default_wasserstein(10.0),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
@@ -255,7 +255,7 @@ end
     end
     @testset ":Max Expectation" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
+        SDDP.adjust_probability(
             default_wasserstein(0.0),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
@@ -266,7 +266,7 @@ end
     end
     @testset ":Min Expectation" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
+        SDDP.adjust_probability(
             default_wasserstein(0.0),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
@@ -277,7 +277,7 @@ end
     end
     @testset ":Max Intermediate" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
+        SDDP.adjust_probability(
             default_wasserstein(0.1),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
@@ -288,7 +288,7 @@ end
     end
     @testset ":Min Intermediate" begin
         risk_adjusted_probability = Vector{Float64}(undef, 4)
-        Kokako.adjust_probability(
+        SDDP.adjust_probability(
             default_wasserstein(0.1),
             risk_adjusted_probability,
             [0.1, 0.2, 0.3, 0.4],
