+++ test/plugins/sampling_schemes.jl
@@ -1,22 +1,22 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-using Kokako, Test
+using SDDP, Test
 
 @testset "InSampleMonteCarlo" begin
     @testset "Acyclic" begin
-        model = Kokako.PolicyGraph(Kokako.LinearGraph(2),
+        model = SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                    lower_bound = 0.0,
                                    direct_mode=false) do node, stage
             @variable(node, 0 <= x <= 1)
-            Kokako.parameterize(node, stage * [1, 3], [0.5, 0.5]) do ω
+            SDDP.parameterize(node, stage * [1, 3], [0.5, 0.5]) do ω
                 JuMP.set_upper_bound(x, ω)
             end
         end
-        scenario, terminated_due_to_cycle = Kokako.sample_scenario(
-            model, Kokako.InSampleMonteCarlo()
+        scenario, terminated_due_to_cycle = SDDP.sample_scenario(
+            model, SDDP.InSampleMonteCarlo()
         )
         @test length(scenario) == 2
         @test !terminated_due_to_cycle
@@ -27,18 +27,18 @@ using Kokako, Test
     end
 
     @testset "Cyclic" begin
-        graph = Kokako.LinearGraph(2)
-        Kokako.add_edge(graph, 2=>1, 0.9)
-        model = Kokako.PolicyGraph(
+        graph = SDDP.LinearGraph(2)
+        SDDP.add_edge(graph, 2=>1, 0.9)
+        model = SDDP.PolicyGraph(
                 graph, lower_bound = 0.0,
                 direct_mode=false) do node, stage
             @variable(node, 0 <= x <= 1)
-            Kokako.parameterize(node, stage * [1, 3], [0.5, 0.5]) do ω
+            SDDP.parameterize(node, stage * [1, 3], [0.5, 0.5]) do ω
                 JuMP.set_upper_bound(x, ω)
             end
         end
-        scenario, terminated_due_to_cycle = Kokako.sample_scenario(
-            model, Kokako.InSampleMonteCarlo(
+        scenario, terminated_due_to_cycle = SDDP.sample_scenario(
+            model, SDDP.InSampleMonteCarlo(
                 terminate_on_dummy_leaf = false,
                 max_depth = 4
             )
