+++ test/user_interface.jl
@@ -1,13 +1,13 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-using Kokako, Test, GLPK
+using SDDP, Test, GLPK
 
 @testset "Basic Graphs" begin
     @testset "LinearGraph" begin
-        graph = Kokako.LinearGraph(5)
+        graph = SDDP.LinearGraph(5)
         @test graph.root_node == 0
         for stage in 0:4
             @test haskey(graph.nodes, stage)
@@ -20,23 +20,23 @@ using Kokako, Test, GLPK
     @testset "MarkovianGraph" begin
         @testset "Error cases" begin
             # Not root transition matrix.
-            @test_throws Exception Kokako.MarkovianGraph([[0.5 0.5; 0.5 0.5]])
+            @test_throws Exception SDDP.MarkovianGraph([[0.5 0.5; 0.5 0.5]])
             # Negative probability.
-            @test_throws Exception Kokako.MarkovianGraph([[-0.5 0.75]])
+            @test_throws Exception SDDP.MarkovianGraph([[-0.5 0.75]])
             # Proability sums to greater than 1.
-            @test_throws Exception Kokako.MarkovianGraph([[0.8 0.8]])
+            @test_throws Exception SDDP.MarkovianGraph([[0.8 0.8]])
             # Mis-matched dimensions.
-            @test_throws Exception Kokako.MarkovianGraph([
+            @test_throws Exception SDDP.MarkovianGraph([
                 [0.1 0.2 0.7], [0.5 0.5; 0.5 0.5]
             ])
         end
         @testset "keyword vs list" begin
-            graph_1 = Kokako.MarkovianGraph(
+            graph_1 = SDDP.MarkovianGraph(
                 stages = 2,
                 transition_matrix = [0.4 0.6; 0.25 0.75],
                 root_node_transition = [0.7, 0.3]
             )
-            graph_2 = Kokako.MarkovianGraph([
+            graph_2 = SDDP.MarkovianGraph([
                 [0.7 0.3], [0.4 0.6; 0.25 0.75]
             ])
             @test graph_1.root_node == graph_2.root_node
@@ -46,65 +46,65 @@ using Kokako, Test, GLPK
 
     @testset "Graph" begin
         @testset "Construct Graph" begin
-            graph = Kokako.Graph(:root)
+            graph = SDDP.Graph(:root)
             @test graph.root_node == :root
             @test collect(keys(graph.nodes)) == [:root]
         end
         @testset "Add node" begin
-            graph = Kokako.Graph(:root)
-            Kokako.add_node(graph, :x)
+            graph = SDDP.Graph(:root)
+            SDDP.add_node(graph, :x)
             @test collect(keys(graph.nodes)) == [:root, :x]
         end
         @testset "Add duplicate node" begin
-            graph = Kokako.Graph(:root)
-            Kokako.add_node(graph, :x)
-            @test_throws Exception Kokako.add_node(graph, :x)
+            graph = SDDP.Graph(:root)
+            SDDP.add_node(graph, :x)
+            @test_throws Exception SDDP.add_node(graph, :x)
         end
         @testset "Add edge" begin
-            graph = Kokako.Graph(:root)
-            Kokako.add_node(graph, :x)
-            Kokako.add_edge(graph, :root => :x, 1.0)
+            graph = SDDP.Graph(:root)
+            SDDP.add_node(graph, :x)
+            SDDP.add_edge(graph, :root => :x, 1.0)
             @test haskey(graph.nodes, :root)
             @test graph.nodes[:root] == [(:x, 1.0)]
         end
         @testset "Add edge of wrong type" begin
-            graph = Kokako.Graph(:root)
-            @test_throws Exception Kokako.add_node(graph, 1)
+            graph = SDDP.Graph(:root)
+            @test_throws Exception SDDP.add_node(graph, 1)
         end
         @testset "Add edge to missing node" begin
-            graph = Kokako.Graph(:root)
-            Kokako.add_node(graph, :x)
-            @test_throws Exception Kokako.add_edge(graph, :x => :y, 1.0)
-            @test_throws Exception Kokako.add_edge(graph, :y => :x, 1.0)
+            graph = SDDP.Graph(:root)
+            SDDP.add_node(graph, :x)
+            @test_throws Exception SDDP.add_edge(graph, :x => :y, 1.0)
+            @test_throws Exception SDDP.add_edge(graph, :y => :x, 1.0)
         end
         @testset "Add edge to root" begin
-            graph = Kokako.Graph(:root)
-            Kokako.add_node(graph, :x)
-            @test_throws Exception Kokako.add_edge(graph, :x => :root, 1.0)
+            graph = SDDP.Graph(:root)
+            SDDP.add_node(graph, :x)
+            @test_throws Exception SDDP.add_edge(graph, :x => :root, 1.0)
         end
         @testset "Invalid probability" begin
-            graph = Kokako.Graph(:root)
-            Kokako.add_node(graph, :x)
-            Kokako.add_edge(graph, :root => :x, 0.5)
-            Kokako.add_edge(graph, :root => :x, 0.75)
-            @test_throws Exception Kokako.validate_graph(graph)
+            graph = SDDP.Graph(:root)
+            SDDP.add_node(graph, :x)
+            SDDP.add_edge(graph, :root => :x, 0.5)
+            SDDP.add_edge(graph, :root => :x, 0.75)
+            @test_throws Exception SDDP.validate_graph(graph)
         end
     end
 end
 
 @testset "PolicyGraph constructor" begin
     @testset "LinearGraph" begin
-        model = Kokako.PolicyGraph(Kokako.LinearGraph(2),
+        model = SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                    lower_bound = 0.0,
                                    direct_mode=false) do node, stage
         end
 
-        @test_throws Exception Kokako.PolicyGraph(Kokako.LinearGraph(2),
+        @test_throws Exception SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                                   lower_bound = 0.0
                                                       ) do node, stage
         end
         nodes = Set{Int}()
-        model = Kokako.PolicyGraph(Kokako.LinearGraph(2),
+        model = SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                    lower_bound = 0.0,
                                    optimizer = with_optimizer(GLPK.Optimizer)
                                        ) do node, stage
@@ -114,7 +114,7 @@ end
     end
 
     @testset "MarkovianGraph" begin
-        graph = Kokako.MarkovianGraph([
+        graph = SDDP.MarkovianGraph([
                 ones(Float64, (1, 1)),
                 [0.5 0.5],
                 [0.5 0.5; 0.3 0.4],
@@ -123,7 +123,7 @@ end
             ]
         )
         nodes = Set{Tuple{Int, Int}}()
-        model = Kokako.PolicyGraph(graph, lower_bound = 0.0,
+        model = SDDP.PolicyGraph(graph, lower_bound = 0.0,
                                    direct_mode = false) do node, stage
             push!(nodes, stage)
         end
@@ -131,7 +131,7 @@ end
     end
 
     @testset "General" begin
-        graph = Kokako.Graph(
+        graph = SDDP.Graph(
             :root,
             [:stage_1, :stage_2, :stage_3],
             [
@@ -142,7 +142,7 @@ end
             ]
         )
         nodes = Set{Symbol}()
-        model = Kokako.PolicyGraph(graph, lower_bound = 0.0,
+        model = SDDP.PolicyGraph(graph, lower_bound = 0.0,
                                    direct_mode = false) do node, stage
             push!(nodes, stage)
         end
@@ -150,11 +150,11 @@ end
     end
 end
 
-@testset "Kokako.State" begin
-    model = Kokako.PolicyGraph(Kokako.LinearGraph(2),
+@testset "SDDP.State" begin
+    model = SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                lower_bound = 0.0,
                                direct_mode = false) do node, stage
-        @variable(node, x, Kokako.State, initial_value = 0)
+        @variable(node, x, SDDP.State, initial_value = 0)
     end
     for stage in 1:2
         node = model[stage]
@@ -164,12 +164,12 @@ end
     end
 end
 
-@testset "Kokako.parameterize" begin
-    model = Kokako.PolicyGraph(Kokako.LinearGraph(2),
+@testset "SDDP.parameterize" begin
+    model = SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                lower_bound = 0.0,
                                direct_mode = false) do node, stage
         @variable(node, 0 <= x <= 1)
-        Kokako.parameterize(node, [1, 2, 3], [0.4, 0.5, 0.1]) do ω
+        SDDP.parameterize(node, [1, 2, 3], [0.4, 0.5, 0.1]) do ω
             JuMP.set_upper_bound(x, ω)
         end
     end
@@ -182,9 +182,9 @@ end
     @test JuMP.upper_bound(node.subproblem[:x]) == 3
 end
 
-@testset "Kokako.set_stage_objective" begin
+@testset "SDDP.set_stage_objective" begin
     @testset ":Min" begin
-        model = Kokako.PolicyGraph(Kokako.LinearGraph(2),
+        model = SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                    lower_bound = 0.0,
                                    direct_mode = false) do node, stage
             @variable(node, 0 <= x <= 1)
@@ -192,11 +192,11 @@ end
         end
         node = model[2]
         @test node.stage_objective == 2 * node.subproblem[:x]
-        @test model.objective_sense == Kokako.MOI.MIN_SENSE
+        @test model.objective_sense == SDDP.MOI.MIN_SENSE
     end
 
     @testset ":Max" begin
-        model = Kokako.PolicyGraph(Kokako.LinearGraph(2),
+        model = SDDP.PolicyGraph(SDDP.LinearGraph(2),
                                    lower_bound = 0.0,
                                    sense = :Max,
                                    direct_mode = false) do node, stage
@@ -205,6 +205,6 @@ end
         end
         node = model[2]
         @test node.stage_objective == 2 * node.subproblem[:x]
-        @test model.objective_sense == Kokako.MOI.MAX_SENSE
+        @test model.objective_sense == SDDP.MOI.MAX_SENSE
     end
 end
