+++ test/visualization.jl
@@ -1,9 +1,9 @@
-#  Copyright 2018, Oscar Dowson.
+#  Copyright 2017-19, Oscar Dowson.
 #  This Source Code Form is subject to the terms of the Mozilla Public
 #  License, v. 2.0. If a copy of the MPL was not distributed with this
 #  file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-using Kokako, Test
+using SDDP, Test
 
 @testset "SpaghettiPlot" begin
     simulations = [
@@ -17,14 +17,14 @@ using Kokako, Test
             Dict(:x => 3.5, :y => 6.5)
         ]
     ]
-    plt = Kokako.SpaghettiPlot(simulations)
-    Kokako.add_spaghetti(plt, cumulative = true) do data
+    plt = SDDP.SpaghettiPlot(simulations)
+    SDDP.add_spaghetti(plt, cumulative = true) do data
         return data[:x]
     end
-    Kokako.add_spaghetti(plt, title = "y") do data
+    SDDP.add_spaghetti(plt, title = "y") do data
         return 2 * data[:y]
     end
-    Kokako.prep_html(plt, "test.html")
+    SDDP.prep_html(plt, "test.html")
     @test read("test.html", String) == read("control.html", String)
     rm("test.html")
 end
